requirements.txt
torch
numpy
mysql-connector-python
pandas
matplotlib
python-dotenv
python
chromadb
langchain
langchain-openai
langchain_huggingFace
langchain-core
transformers
huggingface-hub
mistralai
langsmith
tiktoken
langchain-community
langchain_mistralai
sentence-transformers
pickle
scikit-learn
faiss-cpu
langgraph